require pundit

+++ app/controllers/messages_controller.rb
@@ -1,8 +1,8 @@
 class MessagesController < ApplicationController
-  load_and_authorize_resource except: [:index, :show]
-  authorize_resource only: [:index, :show]
-  before_filter :get_user, only: :index
-  after_filter :solr_commit, only: [:create, :update, :destroy, :destroy_selected]
+  before_action :set_message, only: [:show, :edit, :update, :destroy]
+  before_action :check_policy, only: [:index, :new, :create, :destroy_selected]
+  before_action :get_user, only: :index
+  after_action :solr_commit, only: [:create, :update, :destroy, :destroy_selected]
 
   # GET /messages
   # GET /messages.json
@@ -43,7 +43,6 @@ class MessagesController < ApplicationController
   # GET /messages/1
   # GET /messages/1.json
   def show
-    @message = current_user.received_messages.find(params[:id])
     @message.transition_to!(:read) if @message.current_state != 'read'
 
     respond_to do |format|
@@ -131,7 +130,7 @@ class MessagesController < ApplicationController
       end
       if messages.present?
         messages.each do |message|
-          message.destroy
+          message.destroy if message.receiver == current_user
         end
         flash[:notice] = t('message.messages_were_deleted')
         format.html { redirect_to messages_url }
@@ -143,6 +142,19 @@ class MessagesController < ApplicationController
   end
 
   private
+  def set_message
+    if current_user
+      @message = current_user.received_messages.find(params[:id])
+      authorize @message
+    else
+      access_denied; return
+    end
+  end
+
+  def check_policy
+    authorize Message
+  end
+
   def message_params
     params.require(:message).permit(
       :subject, :body, :sender, :recipient, :parent_id
